Tests: take screenshots much more reliably
`grabWindow()` takes screenshots from X.
This depends on the display and can come with all sorts of artifacts.
Also, it's a problem if another window is overlapping
the one we are interested in.

And `widget.grab()` makes widget render directly to image.
Quite silly of me to not have thought of looking
for something like this in the first place, eh?